requirements.txt
streamlit
langchain-community
langchain-text-splitters
langchain-chroma
chromadb
python-dotenv
sentence-transformers
google-generativeai
